refactoring methods signature modification and moving SubmitPanel in the right package location
