+++ src/main/resources/application.properties
@@ -1,2 +1,2 @@
-#Wed Dec 04 17:00:14 CET 2024
+#Sat Dec 07 16:43:15 CET 2024
 version=0.3.1
